requirements.txt
pandas>=2.0.0
yfinance>=0.2.18
psycopg2-binary>=2.9.6
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
ta>=0.10.2
matplotlib>=3.7.0
numpy>=1.24.0
streamlit>=1.28.0
plotly>=5.17.0
